feat(python): Avoid importing interchange module by default
